+++ lib/navigation/dalal_nav_buidler.dart
@@ -6,7 +6,7 @@ import 'package:dalal_street_client/models/snackbar/snackbar_type.dart';
 import 'package:dalal_street_client/navigation/home_routes.dart';
 import 'package:dalal_street_client/navigation/nav_utils.dart';
 import 'package:dalal_street_client/utils/snackbar.dart';
-import 'package:dalal_street_client/utils/streams_to_toast.dart';
+import 'package:dalal_street_client/utils/stream_snackbar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
